removing (AstarBase) duplicated code
