+++ functions/AstarBase.py
@@ -3,7 +3,7 @@ from pgRoutingLayer import pgRoutingLayer_utils as Utils
 from .FunctionBase import FunctionBase
 from psycopg2 import sql
 
-class Function(FunctionBase):
+class AstarBase(FunctionBase):
 
     minPGRversion = 2.4
 
@@ -33,7 +33,7 @@ class Function(FunctionBase):
             SELECT seq, '(' || start_vid || ',' || end_vid || ')' AS path_name,
                 path_seq AS _path_seq, start_vid AS _start_vid, end_vid AS _end_vid,
                 node AS _node, edge AS _edge, cost AS _cost, lead(agg_cost) over() AS _agg_cost
-            FROM pgr_astar('
+            FROM {function}('
                 {innerQuery}
                 ',
                 {source_ids}, {target_ids}, {directed})
